Change to Spring Boot
Change codebase to work with Spring Boot and use Spring-Boot, Thymeleaf Integration.
