+++ src/main/java/com/climate/climatevisualizer/web/Address.java
@@ -1,0 +1,13 @@
+package com.climate.climatevisualizer.web;
+
+public class Address {
+    private String address;
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+}
