+++ src/main/resources/templates/failedQuery.html
@@ -7,8 +7,7 @@
 </head>
 <body>
     <h1>Result</h1>
-    <p th:text="'Latitude : ' + ${latLon.first}" />
-    <p th:text="'Latitude : ' + ${latLon.second}" />
-    <a href="/index">Try another address!</a>
+    <a href="/index">Sorry! There seems to have been an error processing your
+        address. Please press here to again!</a>
 </body>
 </html>
