+++ src/main/resources/templates/index.html
@@ -7,8 +7,8 @@
 </head>
 <body>
     <h1>Form</h1>
-    <form action="#" th:action="@{/index}" th:object="${index}" method="post">
-        <p>Address: <input type="text" th:field="*{content}" /></p>
+    <form action="#" th:action="@{/index}" th:object="${address}" method="post">
+        <p>Address: <input type="text" th:field="*{address}" /></p>
         <p><input type="submit" value="Submit" />
             <input type="reset" value="Reset"/></p>
     </form>
